feat(email-classifier): Implement comprehensive production email training system
- Add new MacBook optimization modules for email training pipeline
- Create production dataset management and validation scripts
- Enhance email TRM model integration and training orchestration
- Update .gitignore to exclude production email and training format directories
- Add design documentation for real email classifier training system
- Implement checkpoint management and training progress monitoring
- Refactor existing email training infrastructure for production readiness
- Add hardware validation and resource monitoring components
- Create multi-phase training and dataset split validation scripts
Comprehensive system design focuses on leveraging existing components to create a production-ready email classification model with high accuracy and robust training infrastructure.

+++ models/recursive_reasoning/trm.py
@@ -55,7 +55,7 @@ class TinyRecursiveReasoningModel_ACTV1Config(BaseModel):
     halt_max_steps: int
     halt_exploration_prob: float
 
-    forward_dtype: str = "bfloat16"
+    forward_dtype: str = "float32"
 
     # Alexia: added
     mlp_t: bool = False # use mlp on L instead of transformer
